Move find_refs_to_def
