+++ crates/ra_ide/Cargo.toml
@@ -19,6 +19,7 @@ join_to_string = "0.1.3"
 log = "0.4.8"
 rustc-hash = "1.1.0"
 rand = { version = "0.7.3", features = ["small_rng"] }
+# TODO: check if can remove
 once_cell = "1.3.1"
 
 ra_syntax = { path = "../ra_syntax" }
